Enable pure nix builds
With Nix configured to use flakes we can now compile the game in a
controlled environment by running the following:

  nix build

To enable this I took advantage of `bindgenHook` provided by
`rustPlatform` from `nixpkgs`. This also simplifies the development
shell setup. No more messing around with strings in the environment. See
here for more details:

    https://github.com/NixOS/nixpkgs/pull/159899

Unfortunately the game compiled with `nix build` doesn't run for me. I'm
getting the following:

    $ result/bin/run_game
    /build/6rs1g5jhs50dzb7mxlfgcbjlkfn9cvl8-source
    thread 'main' panicked at 'called `Result::unwrap()` on an `Err` value: Os { code: 2, kind: NotFound, message: "No such file or directory" }', run_godot.rs:11:10
    note: run with `RUST_BACKTRACE=1` environment variable to display a backtrace

I can run without issues with `cargo run` or `./target/release/run_game`:

    $ ./target/release/run_game
    /home/tad/Projects/bevy_jam
    Godot Engine v3.4.4.stable.custom_build - https://godotengine.org
    OpenGL ES 2.0 Renderer: AMD RENOIR (LLVM 13.0.1, DRM 3.42, 5.15.59)
    OpenGL ES Batching: ON

It must have something to do with the godot/ directory not being found
at the expected path. I'm not yet sure how it works, but it seems to me
that the path is hardcoded during the compilation (see the first line of
the output). If so, won't this be a problem when we try to distribute
the binary? Or are we ok with people starting the game with `cargo run`?
Or am I confused about something (there's always a chance)?