Write more about the desing
As discussed during video call, I put some more thoughts about the game
design in the readme. I tried to keep it light-hearted, so we could use
it as a promotional material or instructions to players who like to read
a lot.

I'm not sure if we can implement all of this in time for the jam.
Probably we will have to prioritize and get the most important stuff
done and leave some stuff for later. I hope this will help us organize
our work.