When the zombies' population is large, slow the growth
Basically the probability of spawning is an inverse square root of the
population size. So there is no hard limit, but at some point the growth
slows down a lot. Next step: despawn zombies that wander too far from
the player. This two together should keep the population in check.

Also I zoomed out the camera 4x so we can see more area around the player.