Give go to goal a visual indication
The go to goal is now a scene in Godot and an singleton entity in Bevy.
It is indicated with a circle that I took from Feather Icons.

For some reason the player is not moving to the middle of the circle,
but stops when it's edge is touching the middle. I think for now it's
good enough.