Implement aiming and shooting
Press and hold RMB to aim. While aiming the character will stand. Wait
for the character to turn toward the target. Release to shoot.

After shooting, the character will duck. Press R to continue running to
the previously set goal or click with LMB to walk toward new goal.

The code needs some refactoring, but functionally it's working the way I
like it. Maybe the shooting shouldn't be bound to the release of RMB,
but to a dedicated key?